requirements.txt
opencv-python 
pillow 
numpy
pandas
fastai
cv2
pillow-pyheif
gradio